+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning].
 
 - Added Java bindings.
   [#455](https://github.com/ethereum/evmc/pull/455)
+- The C++ EVMC basic types `address` and `bytes32` have all the comparison operators supported.
+  [#474](https://github.com/ethereum/evmc/pull/474)
 
 
 ## [7.1.0] — 2019-11-29